+++ fxkc-modules/fxkc-peis/src/main/java/org/fxkc/peis/domain/bo/TjRegisterBo.java
@@ -11,6 +11,7 @@ import org.fxkc.common.core.validate.EditGroup;
 import org.fxkc.common.mybatis.core.domain.BaseEntity;
 import org.fxkc.peis.domain.TjRegister;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -259,32 +260,32 @@ public class TjRegisterBo extends BaseEntity {
     /**
      * 总费用
      */
-    private Long totalAmount;
+    private BigDecimal totalAmount;
 
     /**
      * 个人费用
      */
-    private Long personAmount;
+    private BigDecimal personAmount;
 
     /**
      * 单位费用
      */
-    private Long teamAmount;
+    private BigDecimal teamAmount;
 
     /**
      * 已缴总费用
      */
-    private Long paidTotalAmount;
+    private BigDecimal paidTotalAmount;
 
     /**
      * 已缴个人费用
      */
-    private Long paidPersonAmount;
+    private BigDecimal paidPersonAmount;
 
     /**
      * 已缴单位费用
      */
-    private Long paidTeamAmount;
+    private BigDecimal paidTeamAmount;
 
     /**
      * 体检登记(报道)时间
